🏗️ (new): Add JSON data file of cities and decode file to display data
Added a JSON file that contains the City data and added a new method in DataService for decoding the JSON data. Also changed the method call in CityView to display the data on screen from the decoded JSON file, instead of using the previous method to retrieve the hard-coded cities data.