✨ (feat): Add game win-lose conditions and countdown timer
Added a countdown timer where the user has to win the game by matching all cards wihtin the allotted time. If there are any un-matched cards left when the timer ends, the player loses the game. An alert popup has been added to display to players whether they've won or lost the game.